Don't retry failed IdV background jobs
**Why**: While testing IdV in staging, we noticed the Sidekiq jobs
were failing due to bugs in the equifax gem. The requests to Equifax
were successfully sent, as were the responses back from Equifax, but
the parsing of the XML response was broken, leading to an exception
in the Sidekiq job.

Each time a request is sent to Equifax, it increments the user's
attempts count, which is limited to a certain number per day.
Since the exception in the job happened after the attempt was already
made, and since Sidekiq is configured to retry failed jobs
automatically, attempt counts kept getting incremented each time the
job retried, unnecessarily penalizing the user. Moreover, exceeding
the allowable limit on Equifax means being locked out of other services
that use Equifax for identity verification.

**How**:
- To disable retries, add a custom Sidekiq middleware to rescue the
exception, then set retries to `false`, then raise the exception so
that we can be notified and see it in the logs.
- To improve the UX, create a new error message for this scenario by
rescuing the exception in the VendorValidatorJob and return an instance
of `Idv::VendorResult` with `job_failed: true` as the error.