Add exception_notification gem
**Why**: To allow us to receive real-time notifications via email
any time an error occurs in production. The gem also has support for
other notification channels, such as Slack. It also captures Sidekiq
errors out of the box.

Using this gem is a lot easier and cheaper than maintaining
a self-hosted version of Sentry or Bugsnag, for example.

Visit the gem's repo for more configuration options:
https://github.com/smartinez87/exception_notification